# Optimize performance of selecting current revision