--- a/GitUI/UserControls/RevisionGrid/RevisionDataGridView.cs
+++ b/GitUI/UserControls/RevisionGrid/RevisionDataGridView.cs
@@ -177,6 +177,16 @@ namespace GitUI.UserControls.RevisionGrid
             }
         }
 
+        // Contains the object Id's that will be selected as soon as they are loaded
+        public HashSet<ObjectId> ToBeSelectedObjectIds { get; set; } = new HashSet<ObjectId>();
+
+        private HashSet<int> ToBeSelectedRowIndexes { get; set; } = new HashSet<int>();
+
+        public bool HasSelection()
+        {
+            return ToBeSelectedObjectIds.Any() || ToBeSelectedRowIndexes.Any() || SelectedRows.Count > 0;
+        }
+
         [CanBeNull]
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Reliability", "CA2002:DoNotLockOnObjectsWithWeakIdentity")]
         [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
@@ -316,6 +326,15 @@ namespace GitUI.UserControls.RevisionGrid
         {
             _graphModel.Add(revision, types);
 
+            if (ToBeSelectedObjectIds.Contains(revision.ObjectId))
+            {
+                lock (_backgroundThread)
+                {
+                    ToBeSelectedObjectIds.Remove(revision.ObjectId);
+                    ToBeSelectedRowIndexes.Add(_graphModel.Count - 1);
+                }
+            }
+
             UpdateVisibleRowRange();
         }
 
@@ -422,6 +441,21 @@ namespace GitUI.UserControls.RevisionGrid
                 else
                 {
                     RowCount = count;
+                    }
+
+                    foreach (int toBeSelectedRowIndex in ToBeSelectedRowIndexes)
+                    {
+                        if (count > toBeSelectedRowIndex)
+                        {
+                            Rows[toBeSelectedRowIndex].Selected = true;
+                            ToBeSelectedRowIndexes.Remove(toBeSelectedRowIndex);
+                            if (CurrentCell == null)
+                            {
+                                CurrentCell = Rows[toBeSelectedRowIndex].Cells[1];
+                            }
+
+                            break;
+                        }
                 }
             }
             finally