Optimizes the code to initially select the correct selected revision. This was done after the complete revision log is loaded. This change selects the commit as soon as the commit to select is loaded. Also, it optimizes the code that searches for the fist parent of the current checkout, if the current checkout is not visible (e.g. in file history).

This Pull Request is on top of pr5473. Seperate pr, because I do not wanted to block the main performance fixes for this. Hope you like this change! On small repo's you probably won't notice any difference/

Changes proposed in this pull request:
- The 'SearchRevision' function is a lot faster. In the linux repo from 6 seconds to 0.3 seconds (depending on how far it needs to search in history). This is only visible when filtering, or in the file history. Both only when the current checkout is filtered out.
- The initial selection is done immediately when this commit is loaded. Instead of waiting for entire list to be loaded. This removes the wait for ~10 seconds on the linux repo. Also, if you started scrolling, the grid scrolled back after 10 seconds to the selected commit. This behaviour is fixed.
- When refreshing (f5) the previous selected commits are remembered and reselected. Multiple select is still possible.
 
What did I do to test the code and ensure quality:
- Tested for about 10 days. Did not extensively tested the SearchRevision code. This code is not modified, only put back after code review in pr5473. I did test it briefly in the file history and it seems to work.

Pros/Cons
+ It is fast!
+ No more sudden scrolling after x seconds
- Code is more complex

Has been tested on (remove any that don't apply):
- GIT 2.18
- Windows 10