@gerhardol 2018-10-07
Works for me. No comments on the code right now.
Seem to fix #5353 

Note: This is a WIP until #5473 is fixed