@spdr870 2018-10-09
Rebased on master