@spdr870 2018-10-10
No reason for the lock on _backgroundEvent. Removed it, but did in in #5473. The code blocks that it was locking never run simultaneous.