@spdr870 2018-10-11
@jbialobr I was able to reproduce what you described on the commit not selected. The commit was correctly selected, only the rowcount was not updated constantly. When you scroll with the mouse, it selected the commit. This also applied to the bisect. I changed this. The rowcount is not updated if the graph is not precached. The graph is precached for 1 page + 150 commits. So this happens. When it now happens, the delay for updating the rowcount is set to every 250ms. Faster is possible, but for me the experience is good. Performance cost is minimal because the invoke is only done when necessary.

The GoToCommit is harder to fix. It is the same behaviour as in the 2.51. You need to be fast with ctrl+shift+g to find the issue! But on the linux repo it is easy to reproduce. This is not fixed yet. For me it is not a big issue. Maybe some indication to the user that the commit is not found. This also happens when the commit is beyond the 200.000 (by default) limit of the log. For me this issue is minor.