@jbialobr 2018-10-11
>The GoToCommit is harder to fix.

After your change I consider it fixed. Before the change, I was not able to go a commit that was returned by git but was not a part of the initial load. Now I am able to do so. Thank you, great work.